feat(ParametricAverage): modify task

+++ week-01/day-4/ParametricAverage/src/ParametricAverage.java
@@ -5,13 +5,13 @@ public class ParametricAverage {
         Scanner sc = new Scanner(System.in);
         System.out.println("How many numbers do you want to handle? ");
         int a = sc.nextInt();
-        double sum = 0;
+        int sum = 0;
         for (int i = 1; i <= a; i++){
             System.out.println("Enter the number nr. " + i + " :");
             double n = sc.nextInt();
             sum += n;
         }
-        double c = sum/a;
+        double c = ((double) sum)/a;
         System.out.println("Sum: " + sum + ", Average: " + c);
     }
 }
